Updates to support Private IP Access to Controller
This change along with a change to the CloudFormation JSON that deploys Controller HA, will allow for the API_PRIVATE_ACCESS environment variable to be maintained properly in the Lambda function.  It will also remove the "public ip" requirement in the Autoscaling Launch Configuration when the controller is accessible over Private IP.

Also, the Print of "0.0.0.0:443/0" message is out of place and is printed even when a modification is no needed.  Relocated to proper function