AVX-60919 Don't disassociate the temporary SG for all Lambda calls (#102)
In cases where the Lambda has to setup a temporary SG during controller
HA, the previous code used to remove the SG if the Lambda gets
retriggered, in an attempt to cleanup after incomplete executions. We
should do this if the Lambda is triggered by (HTTP) function events
(e.g. used to retrieve the software version number).

Update the e2e terraform to use a non-0.0.0.0/0 default SG rule so we
can exercise this path.